Even more base class improvements...
Fixed the GameObject velocity update to be tied to a multiplier, allows
frame-independent movement. Added a set property to the UIBase screen
position, got rid of the abstract methods, added a tag to run on a
variable time scale, added a try block to the content load, and added a
basic Logger class.