Tons of AI additions and bugfixes.
As well as a content update. Finally fixed a vector tracking bug in
GameObject... a little hacky, but it works. Added a level description
field, as well as bugfixes for LevelManager. Major additions to
NPC/Manager, including a shield and better content handling. Polished
the credits screen to flow better, as well as fixing a few double
resets.